refactor: modal layer 변경 widgets -> features
