+++ src/features/comments/index.ts
@@ -1,1 +1,2 @@
 export * from "./model";
+export { CommentAddModal, CommentUpdateModal } from "./ui";
