+++ src/features/comments/ui/index.ts
@@ -1,0 +1,2 @@
+export { CommentAddModal } from "./CommentAddModal";
+export { CommentUpdateModal } from "./CommentUpdateModal";
