+++ src/features/posts/ui/PostDetailModal.tsx
