+++ src/features/posts/ui/index.ts
@@ -1,2 +1,5 @@
 export { PostFilter } from "./PostFilter";
 export { PostTitle } from "./PostTitle";
+export { PostAddModal } from "./PostAddModal";
+export { PostUpdateModal } from "./PostUpdateModal";
+export { PostDetailModal } from "./PostDetailModal";
