+++ src/features/users/ui/index.ts
@@ -1,1 +1,2 @@
 export { UserProfile } from "./UserProfile";
+export { UserProfileModal } from "./UserProfileModal";
