+++ src/widgets/comments/ui/index.ts
@@ -1,3 +1,1 @@
-export { CommentAddModal } from "./CommentAddModal";
-export { CommentUpdateModal } from "./CommentUpdateModal";
 export { Comments } from "./Comments";
