+++ src/widgets/posts/ui/index.ts
@@ -1,4 +1,1 @@
-export { PostAddModal } from "./PostAddModal";
-export { PostUpdateModal } from "./PostUpdateModal";
-export { PostDetailModal } from "./PostDetailModal";
 export { PostTable } from "./PostTable";
