--- src/widgets/users/index.ts
@@ -1,1 +1,0 @@
-export { UserProfileModal } from "./ui";
