--- src/widgets/users/ui/index.ts
@@ -1,1 +1,0 @@
-export { UserProfileModal } from "./UserProfileModal";
